+ internal restructure
