+++ pom.xml
@@ -79,9 +79,6 @@
 				<includes>
 					<include>**/*.yml</include>
 				</includes>
-				<includes>
-					<include>**/*.yml</include>
-				</includes>
 				<filtering>true</filtering>
 			</resource>
 		</resources>
