--- src/main/java/eu/illyrion/illyriacore/interfaces/ItemsInterface.java
@@ -1,16 +1,0 @@
-package eu.illyrion.illyriacore.interfaces;
-
-public interface ItemsInterface {
-    // folder
-    String ITEMS_YML = "items.yml";
-    // properties
-    String ITEM_DURABILITY = "durability";
-    String ITEM_FLAGS = "itemFlags";
-    String ITEM_CUSTOM_MODEL_DATA = "customModelData";
-    String ITEM_UNBREAKABLE = "unbreakable";
-    String ITEM_ENCHANTMENT = "enchantments";
-    String ITEM_LORE = "lores";
-    String ITEM_NAME = "name";
-    String ITEM_MATERIAL = "material";
-    String ITEMS = "items";
-}
