+++ src/main/java/org/xodium/illyriacore/interfaces/PermissionsInterface.java
@@ -1,0 +1,8 @@
+package org.xodium.illyriacore.interfaces;
+
+public interface PermissionsInterface {
+
+    // General
+    public static final String PERM_PREFIX = "illyriautils.";
+    public static final String RELOAD = PERM_PREFIX + "reload";
+}
