--- src/test/java/eu/illyrion/illyriacore/configs/PermissionsTest.java
@@ -1,16 +1,0 @@
-package eu.illyrion.illyriacore.configs;
-
-import org.junit.jupiter.api.Test;
-
-import eu.illyrion.illyriacore.interfaces.PermissionsInterface;
-
-import static org.junit.jupiter.api.Assertions.assertEquals;
-
-public class PermissionsTest implements PermissionsInterface {
-
-    @Test
-    public void testPermissions() {
-        assertEquals("illyriautils.", PERM_PREFIX);
-        assertEquals("illyriautils.reload", RELOAD);
-    }
-}
