--- src/test/java/eu/illyrion/illyriacore/handlers/CustomItemHandlerTest.java
@@ -1,5 +1,0 @@
-package eu.illyrion.illyriacore.handlers;
-
-public class CustomItemHandlerTest {
-
-}
