+++ src/test/java/org/xodium/illyriacore/handlers/CustomItemHandlerTest.java
@@ -1,0 +1,5 @@
+package org.xodium.illyriacore.handlers;
+
+public class CustomItemHandlerTest {
+
+}
